Remove CI tests for versioned releases of Bazel and dependencies.
Unfortunately, the team maintaining this repository does not have the resources to keep the versioned releases up to date. Going forward, the recommended approach for depending on this repository is to select a specific commit and to use the `git_repository` WORKSPACE rule available in `@bazel_tools//tools/build_defs/repo:git.bzl`.

RELNOTES: None
PiperOrigin-RevId: 274255348