Add multi_arch_wrapper for macOS CLIs
This is a rule + transition for forcing macOS CLIs are built with
multiple architectures, even in the host configuration (with `cfg =
"exec"`). This is useful to share caches across Intel and Apple Silicon
machines, and to be able to populate the cache from either machine as
well. This needs to live in apple_support instead of rules_apple so that
rules_swift can use it for the Swift worker.