Add `macos_universal_binary` rule to build multi-arch macOS binaries
This is basically a copy of the new `apple_universal_binary` rule but
with a simpler interface (there are no `minimum_os_version` and
`platform_type` attributes), and the output binary is forced to be
multi-arch. This rule is meant to be used for host/exec tools so that
they are built as universal binaries regardless the architecture of the
execution platform.