Add documentation generation
This currently only generates the rules.md file, but it's a start! This
also updates the manually edited READMEs with new links to these.

This is a bit weird because the apple_genrule that folks use directly is
a macro, so we expose the docs for the underlying rule instead.

Fixes https://github.com/bazelbuild/apple_support/issues/62