Add fixedReleaser to fix BCR app pipeline
The GitHub app to automatically publish new releases to the Bazel Central Registry was failing because it was looking for a BCR fork under the user who tags the release. In our case, this is automated by a GitHub action and so there’s no fork. By adding this config file, we override the default releaser to be my username, so the fork https://github.com/balestrapatrick/bazel-central-registry will be correctly found.