Add Apple CC toolchain setup (#113)
This migrates the Apple CC toolchain from bazel into this repo. This
allows this to evolve and be updated without a full bazel udpate.
Bazel's built in unix toolchain can still be used to build macOS C++
targets, but this repo handles a wider variety of Apple specific flags
and requires you have Xcode installed.