Remove contains_objc_source feature
This feature did 2 things:

1. It passed `-fobjc-link-runtime` to all C++ actions
2. It passed `-framework UIKit` on non macOS platforms to C++ link
   actions

`-fobjc-link-runtime` was also being passed elsewhere to the ObjC link
actions so this flag moved to 1 place in the `default_link_flags` and no
longer needs to be in this feature.

For ObjC link actions we were already passing `-framework UIKit` for
iOS. This makes this feature also apply to C++ link actions.