Add BAZEL_ALLOW_NON_APPLICATIONS_XCODE to run xcode-locator (#197)
Since this toolchain requires Xcode, running xcode_locator during
toolchain setup is likely duplicative. The only things it's used
for is to check that they exist, and to add to the allowed include
directories. The include directories already includes /Applications
which is where most people likely have Xcode installed, so as long
as that can be relied on, fetching these values isn't necessary. This
saves a few seconds when the toolchain is reconfigured.

https://bazelbuild.slack.com/archives/CD3QY5C2X/p1652819625121889