Move -ObjC into a separate feature
This extracts the -ObjC flag into a separate crosstool feature, and
makes it apply to all CC linking actions. Previously this only applied
to objc-executable actions, but as this gets unified into the cc linking
behavior it makes sense to apply it to the cc actions as well. This has
the potential downside that if you had a cc_binary target that had
Objective-C in the dependency tree that it did not use, previously it
would not have been loaded and now it would. If that becomes a problem
this flag can be removed on thew binary in question by disabling this
feature. This also allows us to remove this flag from rules_swift's
logic since it was there for the same reason.