Remove -ObjC from C++ links
Since we went a different route in rules_swift and require the -ObjC
flag to be disabled there too, this isn't as necessary. This reduces the
risk of people making new reliance on -ObjC and when other changes land
to flip the default on this we won't have to worry about that case.