[RFC] Add support for alternate linkers
This is a RFC for adding support for using a different linker than the
default based on a new `--feature`. This is a theoretically replacement
for solutions like [`rules_apple_linker`](https://github.com/keith/rules_apple_linker/).

In some ways this is better, it's easier for users to integrate with,
without a separate rule set, and it's more clear what linker is going to
be used since it's explicitly global vs something in your dependency
tree. Also there is no ambiguity like there is with multiple linkers in
the dependency tree with rules_apple_linker.

This current iteration has some potential UX downsides, like it's harder
to select() on specific linker arguments, so users have to do that
themselves in their macros, but you cannot select() based on features.

Overall I think something like this is worth doing because many users
are interested in at least experimenting with this, and this decreases
the overhead to do so.