Add new XCODE_VERSION toolchain env var (#222)
This can be used to invalidate the toolchain when your Xcode version
changes but the path to Xcode doesn't. This is required if you want to
avoid non-hermetic issues when switching Xcode versions. Previously we
used USE_CLANG_CL for this since it was otherwise ignored, but we should
move to this one which makes it more clear what the purpose is.