Enable -dead_strip by default for opt builds
This change makes `-dead_strip` always be enabled for `opt` builds,
regardless of `--objc_enable_binary_stripping`. This should be the right
default for everyone, if not we can look at adding a way to disable it.

This also allows passing `--features=dead_strip` to enable `-dead_strip`
in other configurations, which previously wasn't supported if you
weren't using `opt`.