Migrate the first batch of Xcode rules' unit tests from Java to Starlark.
This moves tests from `XcodeVersionTest.java` and part of `XcodeConfigTest.java`. The latter is so large that I'm breaking it up into multiple changes to keep the individual ones smaller. It also stubs out the tests from `AvailableXcodesTest.java`, but the provider they use is restricted to built-in Starlark, so we can't actually migrate those yet.

Since the rules in `apple_support` are still just forwarding to the native rules, these tests will be verifying their behavior. This is what we want, because when we migrate the rule implementations themselves over to `apple_support`, we can verify that we did it correctly via these tests.

Unfortunately, some tests can't be migrated, because Starlark tests don't let you set `--experimental_*` flags in config transitions. If those flags are still needed, we may need to migrate them to Starlark as well.

PiperOrigin-RevId: 627111681