Add support for CLT only
The only thing stopping us from supporting the command line tools was
that we always tried to send the Xcode derived environment variables to
every action. For the command line tools there are reasonable
replacements that should always be safe to hardcode. The only thing
weird about this setup is that the target uses the default in bazel
which is currently 10.11. It respects --macos_minimum_os, so that's
probably fine.

In this setup the SDK version is not part of the cache key, users could
pass something in the --action_env to solve this if they wanted. That
seems hard enough to manage just from the CLT installation side so it
probably just doesn't matter for this use case.