Fix bazel @ HEAD compat (#351)
This required adding some loads, and also updating some deps. The
biggest annoyance here was copying the linking logic from rules_apple
since we were doing testing with that relying on bazel itself before. We
might want to revisit this in the future, but IIRC the intent was to
test the toolchain with "real" linking logic so we don't want to just
toss it