Mark module extension as reproducible (#328)
This removes an unnecessary lockfile entry.

The module extension is kept in place so that `use_extension` calls
don't need to be updated. The repo rules are moved into a separate file
so that WORKSPACE setups don't fail on a cyclic load of
`bazel_features`. The newly exposed repo rules are protected by load
visibility.