Added `http_dmg` repository rule for fetching and extracting dmgs (#421)
I've needed this behavior in the past and was able to get something that
works but don't know what the best home would be for such a rule. This
seemed to be the most appropriate place I could find in place of adding.
Looking forward to see if maintainers here agree and if so some
feedback!