Pass -object_path_lto <path> linker flag for LTO builds (#420) (#428)
The following commit (in 1.23.0) was reverted recently (in 1.23.1):
* commit 61b44d12df0834ae26d6e089271e14c6c9a239c8
* revert 4c51f10063687af77eabbbdc035124d798ca3a3f

This patch is a new attempt at getting that feature merged, but without
the LTO related issues that came up in the first patch. The difference
is that we're adding that feature unconditionally now, as it won't hurt
linking actions when LTO is disabled anyway. And enabling the bazel lto
features changes the linking workflow in a way that isn't necessary on
Apple platforms.

Initial commit log below.
---

This is needed for LTO builds, this is the path to which linker writes
one big object file after performing link time optimisations, and then
this object file would be used during final linking to create the App
binary, without this file binary would contain invalid debug symbols,
more info - https://github.com/bazelbuild/rules_swift/issues/1529 .

Closes - https://github.com/bazelbuild/rules_swift/issues/1529

Co-authored-by: Sanju Naik <66404008+sanju-naik@users.noreply.github.com>