Remove unsupported actions from configs
This is just a bit of cleanup until we actually support these actions.
These being here never did anything since we don't register an
action_config for any of these types. If we add these back we should
attempt to add them to one of the existing `all_*_actions` lists
instead, which I plan to make more use of.