fix(xcode_support): handle Xcode version starting with path separator
When Xcode version is passed as a directory, the version check currently
errors out as it obviously cannot compare versions.
We're gonna assume that if the user passes Xcode as a directory, they're
ok managing the features themselves, and we'll assume they're on the
latest and greatest Xcode. So we'll have that logic return True.