Enable code coverage in Xcode for clang-built targets (#491)
This is the first in a stack of PRs across three repositories. 

- bazelbuild/apple_support#491 <-- you are here
- bazelbuild/rules_swift#1623
- MobileNativeFoundation/rules_xcodeproj#3250

It takes the approach suggested in
MobileNativeFoundation/rules_xcodeproj#1119 and applies it to the
xcode-clang toolchain. By breaking sandboxing rules and remapping a
derivable path to the source root, we can to allow Xcode, which
bookkeeps all code coverage representations in the IDE as absolute
paths, to present code coverage information in the gutters and test
results.